Implement Warning the user after successful added if the series was 100 before to reach the maximum 10 digits SeriesNumber in CR

+++ Repository/ReceiptRepo.cs
@@ -13,7 +13,7 @@ namespace Accounting_System.Repository
             _dbContext = dbContext ?? throw new ArgumentNullException(nameof(dbContext));
         }
 
-        public async Task<int> GetLastSeriesNumberCR()
+        public async Task<long> GetLastSeriesNumberCR()
         {
             var lastInvoice = await _dbContext
                 .CollectionReceipts
